@@ -5,44 +5,61 @@ using System;
 using System.Linq;
 using System.Threading.Tasks;
 using BullOak.Repositories.EventStore.Streams;
-using EventStore.Client;
-using StreamNotFoundException = BullOak.Repositories.Exceptions.StreamNotFoundException;
 
 namespace BullOak.Repositories.EventStore
 {
+    using Exceptions;
+
     public class EventStoreRepository<TId, TState> : IStartSessions<TId, TState>, IEventStoreStreamDeleter<TId>
     {
         private readonly IHoldAllConfiguration configs;
-        private readonly EventStoreClient client;
         private readonly IDateTimeProvider dateTimeProvider;
         private readonly IValidateState<TState> stateValidator;
-        private readonly EventReader eventReader;
 
-        private static AlwaysPassValidator<TState> defaultValidator = new AlwaysPassValidator<TState>();
+        private readonly IReadEventsFromStream eventReader;
+        private readonly IStoreEventsToStream eventWriter;
+
+        private static readonly AlwaysPassValidator<TState> defaultValidator = new();
         private static readonly Func<IAmAStoredEvent, bool> alwaysPassPredicate = _ => true;
 
-        public EventStoreRepository(IValidateState<TState> stateValidator, IHoldAllConfiguration configs, EventStoreClient client, IDateTimeProvider dateTimeProvider = null)
+        public EventStoreRepository
+        (
+            IValidateState<TState> stateValidator,
+            IHoldAllConfiguration configs,
+            IReadEventsFromStream eventReader,
+            IStoreEventsToStream eventWriter,
+            IDateTimeProvider dateTimeProvider = null
+        )
         {
             this.stateValidator = stateValidator ?? throw new ArgumentNullException(nameof(stateValidator));
             this.configs = configs ?? throw new ArgumentNullException(nameof(configs));
-            this.eventReader = new EventReader(client, configs);
-            this.client = client ?? throw new ArgumentNullException(nameof(client));
+
+
+            this.eventReader = eventReader;
+            this.eventWriter = eventWriter;
+
             this.dateTimeProvider = dateTimeProvider ?? new SystemDateTimeProvider();
         }
 
-        public EventStoreRepository(IHoldAllConfiguration configs,
-            EventStoreClient client)
-            : this(defaultValidator, configs, client)
-        { }
+        public EventStoreRepository
+        (
+            IHoldAllConfiguration configs,
+            IReadEventsFromStream eventReader,
+            IStoreEventsToStream eventWriter
+        )
+            : this(defaultValidator, configs, eventReader, eventWriter)
+        {
+        }
 
-        public async Task<IManageSessionOf<TState>> BeginSessionFor(TId id, bool throwIfNotExists = false, DateTime? appliesAt = null)
+        public async Task<IManageSessionOf<TState>> BeginSessionFor(TId id, bool throwIfNotExists = false,
+            DateTime? appliesAt = null)
         {
             if (throwIfNotExists && !(await Contains(id)))
                 throw new StreamNotFoundException(id.ToString());
 
-            var session = new EventStoreSession<TState>(stateValidator, configs, client, id.ToString(), dateTimeProvider);
+            var session = new EventStoreSession<TState>(stateValidator, configs, eventReader, eventWriter, id.ToString(), dateTimeProvider);
             await session.Initialize(appliesAt.HasValue
-                ? (IAmAStoredEvent e) => GetBeforeDateEventPredicate(e, appliesAt.Value)
+                ? e => GetBeforeDateEventPredicate(e, appliesAt.Value)
                 : alwaysPassPredicate);
 
             return session;
@@ -50,13 +67,13 @@ namespace BullOak.Repositories.EventStore
 
         private bool GetBeforeDateEventPredicate(IAmAStoredEvent @event, DateTime appliesAt)
         {
-            if (@event.Metadata?.Properties == null || !@event.Metadata.Properties.TryGetValue(MetadataProperties.Timestamp,
-                out var eventTimestamp)) return true;
+            if (@event.Metadata?.Properties == null
+                || !@event.Metadata.Properties.TryGetValue(MetadataProperties.Timestamp, out var eventTimestamp))
+                return true;
 
             if (!DateTime.TryParse(eventTimestamp, out var timestamp)) return true;
 
             return timestamp <= appliesAt;
-
         }
 
         public async Task<bool> Contains(TId selector)
@@ -65,7 +82,7 @@ namespace BullOak.Repositories.EventStore
             {
                 var readResult = await eventReader.ReadFrom(selector.ToString());
 
-                return readResult.streamExists && await readResult.events.AnyAsync();
+                return readResult.StreamExists && await readResult.Events.AnyAsync();
             }
             catch (Exception)
             {
@@ -84,7 +101,8 @@ namespace BullOak.Repositories.EventStore
         /// would consider to be a hard-delete; the scavenger will eventually delete even a soft-deleted stream so if
         /// you want proper soft-delete semantics then use <see cref="SoftDeleteByEvent"/> or
         /// <see cref="SoftDeleteByEvent{TSoftDeleteEventType}"/>
-        [Obsolete("Please use either IEventStoreStreamDeleter.SoftDelete or IEventStoreStreamDeleter.SoftDeleteByEvent")]
+        [Obsolete(
+            "Please use either IEventStoreStreamDeleter.SoftDelete or IEventStoreStreamDeleter.SoftDeleteByEvent")]
         public async Task Delete(TId selector)
         {
             await SoftDelete(selector);
@@ -95,32 +113,8 @@ namespace BullOak.Repositories.EventStore
         {
             var readResult = await eventReader.ReadFrom(selector.ToString());
 
-            if (readResult.streamExists)
-                await client.SoftDeleteAsync(selector.ToString(), StreamState.Any);
-        }
-
-        //[Obsolete("This doesn't care for other operations. In the future it will be moved to session or renamed.")]
-        public Task SoftDeleteByEvent(TId selector)
-            => SoftDeleteByEventImpl(selector, DefaultSoftDeleteEvent.ItemWithType.CreateEventData(dateTimeProvider));
-
-        //[Obsolete("This doesn't care for other operations. In the future it will be moved to session or renamed.")]
-        public async Task SoftDeleteByEvent<TSoftDeleteEventType>(TId selector,
-            Func<TSoftDeleteEventType> createSoftDeleteEventFunc)
-            where TSoftDeleteEventType : EntitySoftDeleted
-        {
-            if (createSoftDeleteEventFunc == null) throw new ArgumentNullException(nameof(createSoftDeleteEventFunc));
-
-            await SoftDeleteByEventImpl(selector, new ItemWithType(createSoftDeleteEventFunc()).CreateEventData(dateTimeProvider));
-        }
-
-        private async Task SoftDeleteByEventImpl(TId selector, EventData softDeleteEvent)
-        {
-            var readResult = await eventReader.ReadFrom(selector.ToString());
-
-            if (readResult.streamExists)
-            {
-                await client.AppendToStreamAsync(selector.ToString(), StreamState.Any, new[] {softDeleteEvent});
-            }
+            if (readResult.StreamExists)
+                await eventWriter.SoftDelete(selector.ToString());
         }
     }
 }
