@@ -4,6 +4,7 @@
     using Session;
     using System;
     using System.Collections.Generic;
+    using Streams;
 
     internal class TestDataContext
     {
@@ -15,10 +16,10 @@
 
         public Exception RecordedException { get; internal set; }
         public IHoldHigherOrder LatestLoadedState { get; internal set; }
-        public Dictionary<string, IManageSessionOf<IHoldHigherOrder>> NamedSessions { get; internal set; } =
-            new Dictionary<string, IManageSessionOf<IHoldHigherOrder>>();
-        public Dictionary<string, List<Exception>> NamedSessionsExceptions { get; internal set; } =
-            new Dictionary<string, List<Exception>>();
+
+        public StreamReadResults LatestStreamReadResults { get; internal set; }
+        public Dictionary<string, IManageSessionOf<IHoldHigherOrder>> NamedSessions { get; } = new();
+        public Dictionary<string, List<Exception>> NamedSessionsExceptions { get; } = new();
 
         public int LastConcurrencyId { get; set; }
 
