@@ -1,13 +1,14 @@
-﻿using EventStore.Client;
-using System;
-using BullOak.Repositories.EventStore.Events;
-using BullOak.Repositories.EventStore.Metadata;
-using Newtonsoft.Json.Linq;
-using BullOak.Repositories.StateEmit;
-
-
-namespace BullOak.Repositories.EventStore
+﻿namespace BullOak.Repositories.EventStore
 {
+    using System;
+    using System.Globalization;
+    using Events;
+    using Metadata;
+    using Newtonsoft.Json.Linq;
+    using StateEmit;
+    using EsClientV20 = global::EventStore.Client;
+    using EsClientV5 = global::EventStore.ClientAPI;
+
     public static class ItemWithTypeExtensions
     {
         private static readonly string CanEditJsonFieldName;
@@ -19,7 +20,7 @@ namespace BullOak.Repositories.EventStore
                                    + CanEditJsonFieldName.Substring(1);
         }
 
-        public static EventData CreateEventData(this ItemWithType @event, IDateTimeProvider dateTimeProvider)
+        public static EsClientV20.EventData CreateV20EventData(this ItemWithType @event, IDateTimeProvider dateTimeProvider)
         {
             var metadata = EventMetadata_V2.From(@event,
                 (MetadataProperties.Timestamp, dateTimeProvider.UtcNow.ToString("yyyy-MM-dd HH:mm:ss")));
@@ -27,9 +28,25 @@ namespace BullOak.Repositories.EventStore
             var eventAsJson = JObject.FromObject(@event.instance);
             eventAsJson.Remove(CanEditJsonFieldName);
 
-            return new EventData(
-                Uuid.NewUuid(),
+            return new EsClientV20.EventData(
+                EsClientV20.Uuid.NewUuid(),
                 @event.type.Name,
+                System.Text.Encoding.UTF8.GetBytes(eventAsJson.ToString()),
+                MetadataSerializer.Serialize(metadata));
+        }
+
+        public static EsClientV5.EventData CreateV5EventData(this ItemWithType @event, IDateTimeProvider dateTimeProvider)
+        {
+            var metadata = EventMetadata_V2.From(@event,
+                (MetadataProperties.Timestamp, dateTimeProvider.UtcNow.ToString("yyyy-MM-dd HH:mm:ss")));
+
+            var eventAsJson = JObject.FromObject(@event.instance);
+            eventAsJson.Remove(CanEditJsonFieldName);
+
+            return new EsClientV5.EventData(
+                Guid.NewGuid(),
+                @event.type.Name,
+                true,
                 System.Text.Encoding.UTF8.GetBytes(eventAsJson.ToString()),
                 MetadataSerializer.Serialize(metadata));
         }
