@@ -1,18 +1,17 @@
 ﻿using System.Collections.Generic;
 using System.Threading;
 using BullOak.Repositories.EventStore.Events;
-using EventStore.Client;
 
 namespace BullOak.Repositories.EventStore.Streams
 {
     using System;
     using System.Threading.Tasks;
 
-    internal interface IReadEventsFromStream
+    public interface IReadEventsFromStream
     {
         Task<StreamReadResults> ReadFrom(string streamId,
             Func<IAmAStoredEvent, bool> predicate = null,
-            Direction direction = Direction.Backwards,
+            StreamReadDirection direction = StreamReadDirection.Backwards,
             CancellationToken cancellationToken = default);
     }
 }
